adding property_set and updating Concepts
properties are an extension point so they can be in a class or applied by external trait specialization.

public:
  using properties - property_set<receiver_tag, single_tag, flow_tag, time_tag>;

Each property is also an extension point so that the category can be inside a property type or applied externally by specialization.

property_set itself is a meta type that should be built on a meta ;) library. more work is needed to make manipulation simpler.

concepts are added to query the properties for an object.

PropertyQuery<>
PropertySetQuery<>
PropertyCategoryQuery<>
PropertySetCategoryQuery<>